requirements.txt
kafka-python
python-twitter
tweepy
pyspark